requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-decouple==3.4
requests==2.25.1
soupsieve==2.1
urllib3==1.26.2
